migrate profile page to next js
