+++ src/app/[[...slug]]/page.jsx
@@ -2,7 +2,7 @@ import '../../index.css'
 import { ClientOnly } from './client'
  
 export function generateStaticParams() {
-  return [{ slug: [''] } , {slug: ['blog']}]
+  return [{ slug: [''] } , {slug: ['blog']}, {slug: ['profile']}]
 }
  
 export default function Page() {
